Feature/delivery address

See merge request leson131002/locker!12

+++ LockerService.Application/Settings/Commands/OrderSettingsCommand.cs
@@ -27,7 +27,7 @@ public class OrderSettingsCommand
 
     public decimal StoragePrice { get; set; }
 
-    public float MaxTimeInHours { get; set; }
+    public int MaxTimeInHours { get; set; }
 
     public decimal ExtraFee { get; set; }
